@@ -1,6 +1,6 @@
 {
   "name": "www-next",
-  "version": "0.2.0",
+  "version": "0.3.0",
   "private": true,
   "scripts": {
     "build": "next lint && next build",
@@ -17,30 +17,30 @@
     "@tailwindcss/typography": "^0.5.4",
     "date-fns": "^2.29.1",
     "next": "^12.2.3",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "tailwindcss": "^3.1.7"
   },
   "devDependencies": {
-    "@jsdevtools/rehype-toc": "^3.0.2",
-    "autoprefixer": "^10.4.0",
-    "eslint": "^8.4.1",
-    "eslint-config-next": "^12.0.7",
+    "autoprefixer": "^10.4.8",
+    "eslint": "^8.21.0",
+    "eslint-config-next": "^12.2.3",
     "github-slugger": "^1.4.0",
     "globby": "^11.0.0",
     "gray-matter": "^4.0.3",
-    "moviedb-promise": "^3.1.12",
-    "next-mdx-remote": "^3.0.8",
-    "postcss": "^8.4.4",
-    "prettier": "^2.5.1",
-    "rehype-autolink-headings": "^6.0.0",
-    "rehype-highlight": "^4.1.0",
-    "rehype-katex": "^5.0.0",
-    "rehype-slug": "^4.0.1",
-    "remark-emoji": "^2.2.0",
+    "moviedb-promise": "^3.1.15",
+    "next-mdx-remote": "^4.1.0",
+    "postcss": "^8.4.14",
+    "prettier": "^2.7.1",
+    "rehype-autolink-headings": "^6.1.1",
+    "rehype-highlight": "^5.0.2",
+    "rehype-katex": "^6.0.2",
+    "rehype-raw": "^6.1.1",
+    "rehype-slug": "^5.0.1",
+    "remark-emoji": "^3.0.2",
     "remark-external-links": "^8.0.0",
-    "remark-footnotes": "^2.0.0",
+    "remark-gfm": "^3.0.1",
     "remark-hint": "^1.0.10",
-    "remark-math": "^3.0.0"
+    "remark-math": "^5.1.1"
   }
 }
